Move global directive to ImageItem component
