+++ src/main.js
@@ -8,20 +8,12 @@ import { valuesI18n } from "./i18n";
 // createI18n: pass the values
 const i18n = createI18n(valuesI18n);
 
-
 // Create the app
 const app = createApp(App);
-
-
-// Imports lazy load directive
-import LazyLoadDirective from "./directives/LazyLoadDirective";
-app.directive("lazyload", LazyLoadDirective);
-
 
 // Add uses
 app.use(router);
 app.use(i18n);
-
 
 // Mounted
 app.mount("#app");
